+++ tests/qiskit_convert_test.py
@@ -762,6 +762,12 @@ def test_qcontrolbox_conversion() -> None:
     assert tkc2.n_gates_of_type(OpType.QControlBox) == 3
 
 
+# Ensures that the tk_to_qiskit converter does not cancel redundant gates
+def test_tk_to_qiskit_redundancies() -> None:
+    h_circ = Circuit(1).H(0).H(0)
+    qc_h = tk_to_qiskit(h_circ)
+    assert qc_h.count_ops()["h"] == 2
+
 def test_state_prep_conversion() -> None:
     # State prep with list of real amplitudes
     ghz_state = [1 / np.sqrt(2), 0, 0, 0, 0, 0, 0, 1 / np.sqrt(2)]