+++ docs/changelog.rst
@@ -2,6 +2,8 @@ Changelog
 ~~~~~~~~~
 0.40.0 (unreleased)
 -------------------
+
+* Fix to the `tk_to_qiskit` converter to prevent cancellation of redundant gates when converting to qiskit.
 * Handle qiskit state initialisation circuits with :py:class:`StatePreparationBox` in the circuit converters.
 
 0.39.0 (May 2023)